requirements.txt
# Telegram бот
python-telegram-bot==21.7

# LLM (Gemini)
google-generativeai==0.8.3

# Векторная база
chromadb==0.5.18
sentence-transformers==3.3.0

# Конфиг
python-dotenv==1.0.1
